requirements.txt
fastapi
uvicorn
pandas
scikit-learn
requests
python-dotenv
